Modify version number
Because of trailing error.

Upgrade "my-app-multi-version" failed: failed to create resource: UIPlugin.catalog.cattle.io "my-app-multi-version" is invalid: metadata.labels: Invalid value: "103.0.0+up1.0.0": a valid label must be an empty string or consist of alphanumeric characters, '-', '_' or '.', and must start and end with an alphanumeric character (e.g. 'MyValue',  or 'my_value',  or '12345', regex used for validation is '(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?')

Signed-off-by: Masashi Honma <masashi.honma@gmail.com>